+++ src/features/events/forms/fields/index.ts
@@ -1,3 +1,0 @@
-export { BasicInfoFields } from './BasicInfoFields'
-export { DescriptionField } from './DescriptionField'
-export { BannerUploadField } from './BannerUploadField'